requirements.txt
accelerate==1.10.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.6.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
distlib==0.3.6
Django==4.1.8
django-bootstrap-v5==1.0.11
django-cors-headers==3.14.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
fastapi==0.116.1
filelock==3.12.0
fonttools==4.59.0
fsspec==2025.9.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==1.0.0rc1
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.5
mpmath==1.3.0
mysqlclient==2.2.7
networkx==3.5
numpy==2.3.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.2
pillow==11.2.1
platformdirs==3.5.0
psutil==7.1.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pygame==2.6.1
pymongo==3.12.3
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2023.3
PyYAML==6.0.2
regex==2025.9.18
requests==2.32.3
safetensors==0.6.2
selenium==4.33.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.2.4
starlette==0.47.2
sympy==1.14.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typer-slim==0.19.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2023.3
urllib3==2.4.0
uvicorn==0.35.0
virtualenv==20.23.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
